librbd/cache: Refactor common code for RWL and SSD
... from AbstractWriteLog class
